move get links properties map helper
