+++ lib/shared/helpers/page-builder/get-links-properties-map.js
