+++ lib/shared/screens/admin/shared/components/page-builder-menu/tabs/link/index.js
@@ -2,7 +2,7 @@ import * as pbActions from 'actions/page-builder';
 import Component from 'components/component';
 import bind from 'decorators/bind';
 import elements from 'elements';
-import getLinksPropertiesMap from 'helpers/get-links-properties-map';
+import getLinksPropertiesMap from 'helpers/page-builder/get-links-properties-map';
 import isContextEqual from 'helpers/page-builder/is-context-equal';
 import React, {PropTypes} from 'react';
 import {bindActionCreators} from 'redux';
